version

+++ docker/core/mkwebaxum/src/admin/bp_database.rs
@@ -24,7 +24,7 @@ struct AdminDBStatsTemplate<'a> {
     template_data_db_size: &'a Vec<mk_lib_database::mk_lib_database_postgresql::PGTableSize>,
     template_data_db_size_total: &'a i64,
     template_data_db_count: &'a Vec<mk_lib_database::mk_lib_database_postgresql::PGTableRows>,
-    template_data_db_count_total: &'a f32,
+    template_data_db_count_total: &'a f64,
     template_data_db_workers: &'a String,
     template_data_db_extension:
         &'a Vec<mk_lib_database::mk_lib_database_postgresql::PGExtensionActive>,
